requirements.txt
scipy==1.11.4
pillow==10.2.0
tensorflow==2.15.0
opencv-python==4.9.0.80
